Rename files to remove cuda suffix
Lets namespace with the packagename, dask_cuda, instead
